+++ src/main/kotlin/racing/ResultView.kt
@@ -1,0 +1,13 @@
+package racing
+
+class ResultView {
+    fun view(carList: List<Car>) {
+        carList.forEach {
+            repeat(it.position) {
+                print("-")
+            }
+            println()
+        }
+        println()
+    }
+}
